input = obj
